requirements.txt
argcomplete==1.11.1
boto3==1.14.4
botocore==1.17.4
certifi==2020.4.5.2
cfn-flip==1.2.3
chardet==3.0.4
click==7.1.2
docutils==0.15.2
durationpy==0.5
Flask==1.1.2
flask-dynamo==0.1.2
future==0.18.2
hjson==3.0.1
idna==2.9
importlib-metadata==1.6.1
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
kappa==0.6.0
MarkupSafe==1.1.1
pip-tools==5.2.1
placebo==0.9.0
python-dateutil==2.6.1
python-dotenv==0.13.0
python-slugify==4.0.0
PyYAML==5.3.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
text-unidecode==1.3
toml==0.10.1
tqdm==4.46.1
troposphere==2.6.1
urllib3==1.25.9
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
zappa==0.51.0
zipp==3.1.0
